Move all utils used once to their app; add lodash-es; add bundle analyzer
