--- libs/utils/src/hash.ts
@@ -1,7 +1,0 @@
-import crypto from "crypto";
-
-export const hash = (source: string) =>
-  crypto
-    .createHash("md5")
-    .update(source)
-    .digest("hex");
