--- libs/utils/src/string.ts
@@ -1,3 +1,0 @@
-export const capitalize = (s: string) => {
-  return s[0].toUpperCase() + s.slice(1).toLowerCase();
-};
