+++ libs/utils/src/url.ts
@@ -1,8 +1,5 @@
 import { IncomingMessage } from "http";
 
-export const normalizeParam = (param: string) =>
-  param.replace(/[\W_]+/g, "_").toLowerCase();
-
 export const urlFromReq = (req: IncomingMessage) => {
   const host = req.headers.host ?? "localhost:3000";
   const protocol = host.split(":")[0] === "localhost" ? "http" : "https";
