+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "analyze": "./tools/scripts/analyze.sh",
     "dev": "./tools/scripts/dev.sh",
     "postinstall": "patch-package && tsc --build --incremental"
   },
@@ -12,11 +13,10 @@
     "aws-sdk": "^2.647.0",
     "csvtojson": "^2.0.10",
     "date-fns": "^2.12.0",
-    "debounce": "^1.2.0",
     "fast-memoize": "^2.5.2",
     "isomorphic-unfetch": "^3.0.0",
     "isotopes": "^0.6.0",
-    "lodash.groupby": "^4.6.0",
+    "lodash-es": "^4.17.15",
     "next": "^9.3.5",
     "next-mdx-enhanced": "^2.4.0",
     "node-cookie": "^2.1.2",
@@ -27,11 +27,11 @@
     "swr": "^0.1.18"
   },
   "devDependencies": {
+    "@next/bundle-analyzer": "^9.3.5",
     "@now/node": "^1.5.0",
     "@types/cookie": "^0.3.3",
     "@types/csvtojson": "^1.1.5",
-    "@types/debounce": "^1.2.0",
-    "@types/lodash.groupby": "^4.6.6",
+    "@types/lodash-es": "^4.17.3",
     "@types/mdx-js__react": "^1.5.1",
     "@types/node-fetch": "^2.5.5",
     "@types/react": "^16.9.23",
@@ -44,6 +44,7 @@
     "dotenv-cli": "^3.1.0",
     "find-up": "^4.1.0",
     "hasura-cli": "^1.1.0",
+    "next-transpile-modules": "^3.2.0",
     "p-limit": "^2.2.2",
     "patch-package": "^6.2.2",
     "styled-components": "4",
